add api for upload file and confirmation success upload

+++ migrations/2021-06-29-101119_ref_table/up.sql
@@ -23,7 +23,7 @@ CREATE TABLE file_ref (
   filename VARCHAR(225) NOT NULL, /* наименование файла */
   content_type VARCHAR(50) NOT NULL, /* тип контента в файле */
   id_ext INTEGER NOT NULL, /* расширение файла (используется для определения CAD) */
-  filesize INTEGER NOT NULL, /* размер файла */
+  filesize BIGINT NOT NULL, /* размер файла */
   path_file VARCHAR(225) NOT NULL, /* путь к файлу */
   created_at TIMESTAMP NOT NULL DEFAULT NOW(), /* дата создания/загрузки */
   updated_at TIMESTAMP NOT NULL DEFAULT NOW(), /* дата обновления */
